requirements.txt
requests==2.31.0
PyYAML==6.0
pytz==2023.3
influxdb-client==1.36.0
urllib3==2.0.3


# pip install -r requirements.txt
